+++ ElementX/Sources/Screens/HomeScreen/View/HomeScreen.swift
@@ -74,6 +74,7 @@ struct HomeScreen: View {
                     }
                 }
                 .searchable(text: $context.searchQuery)
+                .searchableStyle(.list)
                 .disableAutocorrection(true)
             }
         }
@@ -101,6 +102,12 @@ struct HomeScreen: View {
         .toolbar {
             ToolbarItem(placement: .navigationBarLeading) {
                 userMenuButton
+            }
+            if context.viewState.showStartChatFlowEnabled {
+                ToolbarItemGroup(placement: .bottomBar) {
+                    Spacer()
+                    newRoomButton
+                }
             }
         }
         .background(Color.element.background.ignoresSafeArea())
@@ -148,6 +155,12 @@ struct HomeScreen: View {
         .accessibilityLabel(ElementL10n.a11yAllChatsUserAvatarMenu)
     }
     
+    private var newRoomButton: some View {
+        Button(action: startChat) {
+            Image(systemName: "square.and.pencil")
+        }
+    }
+    
     private var sessionVerificationBanner: some View {
         VStack(alignment: .leading, spacing: 16) {
             VStack(alignment: .leading, spacing: 4) {
@@ -191,6 +204,10 @@ struct HomeScreen: View {
         context.send(viewAction: .userMenu(action: .inviteFriends))
     }
 
+    private func startChat() {
+        context.send(viewAction: .startChat)
+    }
+    
     private func feedback() {
         context.send(viewAction: .userMenu(action: .feedback))
     }
@@ -224,9 +241,7 @@ struct HomeScreen: View {
 struct HomeScreen_Previews: PreviewProvider {
     static var previews: some View {
         body(.loading)
-            .tint(.element.accent)
         body(.loaded)
-            .tint(.element.accent)
     }
     
     static func body(_ state: MockRoomSummaryProviderState) -> some View {