+++ ElementX/Sources/Screens/HomeScreen/View/HomeScreenRoomCell.swift
@@ -17,10 +17,14 @@
 import SwiftUI
 
 struct HomeScreenRoomCell: View {
+    @Environment(\.dynamicTypeSize) var dynamicTypeSize
+    @Environment(\.redactionReasons) private var redactionReasons
+    
     let room: HomeScreenRoom
     let context: HomeScreenViewModel.Context
     
-    @Environment(\.dynamicTypeSize) var dynamicTypeSize
+    private let verticalInsets = 12.0
+    private let horizontalInsets = 16.0
     
     var body: some View {
         Button {
@@ -31,22 +35,20 @@ struct HomeScreenRoomCell: View {
             HStack(spacing: 16.0) {
                 avatar
                 
-                VStack(alignment: .leading, spacing: 2) {
-                    header
-                    footer
-                }
+                content
+                    .padding(.vertical, verticalInsets)
+                    .overlay(alignment: .bottom) {
+                        Rectangle()
+                            .fill(Color.element.quinaryContent)
+                            .frame(height: 1 / UIScreen.main.scale)
+                            .padding(.trailing, -horizontalInsets)
+                    }
             }
-            .frame(minHeight: 84.0)
+            .padding(.horizontal, horizontalInsets)
             .accessibilityElement(children: .combine)
         }
-        .buttonStyle(HomeScreenRoomCellButtonStyle())
+        .buttonStyle(HomeScreenRoomCellButtonStyle(isSelected: false))
         .accessibilityIdentifier(A11yIdentifiers.homeScreen.roomName(room.name))
-        .overlay(alignment: .bottom) {
-            Divider()
-                .frame(height: 0.5)
-                .background(Color.element.quinaryContent)
-                .padding(.leading, 84)
-        }
     }
     
     @ViewBuilder
@@ -59,6 +61,23 @@ struct HomeScreenRoomCell: View {
                                 imageProvider: context.imageProvider)
                 .dynamicTypeSize(dynamicTypeSize < .accessibility1 ? dynamicTypeSize : .accessibility1)
                 .accessibilityHidden(true)
+        }
+    }
+    
+    var content: some View {
+        VStack(alignment: .leading, spacing: 2) {
+            header
+            footer
+        }
+        // Hide the normal content for Skeletons and overlay centre aligned placeholders.
+        .opacity(redactionReasons.contains(.placeholder) ? 0 : 1)
+        .overlay {
+            if redactionReasons.contains(.placeholder) {
+                VStack(alignment: .leading, spacing: 2) {
+                    header
+                    lastMessage
+                }
+            }
         }
     }
     
@@ -84,44 +103,53 @@ struct HomeScreenRoomCell: View {
         HStack(alignment: .firstTextBaseline) {
             ZStack(alignment: .topLeading) {
                 // Hidden text with 2 lines to maintain consistent height, scaling with dynamic text.
-                Text(" \n ").lastMessageFormatting().hidden()
+                Text(" \n ")
+                    .lastMessageFormatting()
+                    .hidden()
+                    .environment(\.redactionReasons, []) // Always maintain consistent height
                 
-                switch room.lastMessage {
-                case .loaded(let lastMessage):
-                    Text(lastMessage)
-                        .lastMessageFormatting()
-                case .loading:
-                    Text(HomeScreenRoom.placeholderLastMessage)
-                        .lastMessageFormatting()
-                        .redacted(reason: .placeholder)
-                case .unknown:
-                    Text(ElementL10n.message)
-                        .lastMessageFormatting()
-                }
+                lastMessage
             }
             
             Spacer()
             
             if room.hasUnreads {
-                Rectangle()
+                Circle()
                     .frame(width: 12, height: 12)
                     .foregroundColor(.element.brand)
-                    .clipShape(Circle())
                     .padding(.leading, 12)
             } else {
                 // Force extra padding between last message text and the right border of the screen if there is no unread dot
-                Rectangle()
-                    .fill(Color.clear)
+                Circle()
                     .frame(width: 12, height: 12)
+                    .hidden()
             }
+        }
+    }
+    
+    @ViewBuilder
+    var lastMessage: some View {
+        switch room.lastMessage {
+        case .loaded(let lastMessage):
+            Text(lastMessage)
+                .lastMessageFormatting()
+        case .loading:
+            Text(HomeScreenRoom.placeholderLastMessage)
+                .lastMessageFormatting()
+                .redacted(reason: .placeholder)
+        case .unknown:
+            Text(ElementL10n.message)
+                .lastMessageFormatting()
         }
     }
 }
 
 struct HomeScreenRoomCellButtonStyle: ButtonStyle {
+    let isSelected: Bool
+    
     func makeBody(configuration: Configuration) -> some View {
         configuration.label
-            .roomCellBackground(configuration.isPressed ? .element.system : .clear)
+            .background(configuration.isPressed || isSelected ? Color.element.system : .clear)
             .contentShape(Rectangle())
     }
 }
@@ -133,21 +161,10 @@ private extension View {
             .lineLimit(2)
             .multilineTextAlignment(.leading)
     }
-    
-    // To be used to indicate the selected room too
-    func roomCellBackground(_ background: Color) -> some View {
-        padding(.horizontal, 8)
-            .padding(.horizontal, 8)
-            .background { background }
-    }
 }
 
 struct HomeScreenRoomCell_Previews: PreviewProvider {
     static var previews: some View {
-        body.tint(.element.accent)
-    }
-
-    static var body: some View {
         let summaryProvider = MockRoomSummaryProvider(state: .loaded)
 
         let userSession = MockUserSession(clientProxy: MockClientProxy(userID: "John Doe", roomSummaryProvider: summaryProvider),
@@ -175,6 +192,13 @@ struct HomeScreenRoomCell_Previews: PreviewProvider {
             ForEach(rooms) { room in
                 HomeScreenRoomCell(room: room, context: viewModel.context)
             }
+            
+            HomeScreenRoomCell(room: .placeholder(), context: viewModel.context)
+                .redacted(reason: .placeholder)
+            HomeScreenRoomCell(room: .placeholder(), context: viewModel.context)
+                .redacted(reason: .placeholder)
+            HomeScreenRoomCell(room: .placeholder(), context: viewModel.context)
+                .redacted(reason: .placeholder)
         }
     }
 }