+++ ElementX/Sources/Screens/SessionVerification/View/SessionVerificationScreen.swift
@@ -226,19 +226,16 @@ struct SessionVerificationScreen: View {
 
 struct SessionVerification_Previews: PreviewProvider {
     static var previews: some View {
-        Group {
-            sessionVerificationScreen(state: .initial)
-            sessionVerificationScreen(state: .requestingVerification)
-            sessionVerificationScreen(state: .cancelled)
-            
-            sessionVerificationScreen(state: .showingChallenge(emojis: MockSessionVerificationControllerProxy.emojis))
-            sessionVerificationScreen(state: .verified)
-        }
-        .tint(Color.element.accent)
+        sessionVerificationScreen(state: .initial)
+        sessionVerificationScreen(state: .requestingVerification)
+        sessionVerificationScreen(state: .cancelled)
+        
+        sessionVerificationScreen(state: .showingChallenge(emojis: SessionVerificationControllerProxyMock.emojis))
+        sessionVerificationScreen(state: .verified)
     }
     
     static func sessionVerificationScreen(state: SessionVerificationStateMachine.State) -> some View {
-        let viewModel = SessionVerificationViewModel(sessionVerificationControllerProxy: MockSessionVerificationControllerProxy(),
+        let viewModel = SessionVerificationViewModel(sessionVerificationControllerProxy: SessionVerificationControllerProxyMock.configureMock(),
                                                      initialState: SessionVerificationViewState(verificationState: state))
         
         return SessionVerificationScreen(context: viewModel.context)